Initial commit for v2
Full rewrite as usbselfserial
