--- driver/__init__.py
@@ -1,4 +1,0 @@
-#!/usr/bin/env python
-"""
-usbpty drivers!
-"""
